move get_db_version to cli utils
